+++ src/DataTransfer/DataTransfer/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 ## Upcoming Release
 * General availability for module Az.DataTransfer
+* Added commands for flow management
 
 ## Version 0.1.0
 * First preview release for module Az.DataTransfer