+++ src/DataTransfer/DataTransfer/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* General availability for module Az.DataTransfer
+* Added commands for flow management
 
 ## Version 0.1.0
 * First preview release for module Az.DataTransfer